+++ src/client/src/api/apiInterfaces.ts
@@ -61,6 +61,7 @@ export interface User {
   firstName: string;
   lastName: string;
   isAdmin: boolean;
+  deletable: boolean;
   isDisabled?: boolean;
   disabledAt?: Date | string;
   createdAt?: Date | string;