+++ src/client/src/pages/settings/admin/userDetail.tsx
@@ -1,6 +1,6 @@
-import { ChangeEvent, useCallback, useContext, useEffect, useState } from "react";
+import { ChangeEvent, FC, useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { useParams } from "react-router-dom";
+import { useHistory, useParams } from "react-router-dom";
 import { Card, CardContent, CardHeader, Checkbox, Chip, Stack, Typography } from "@mui/material";
 import { DataGrid, GridColDef, GridFilterModel, GridRenderCellParams, GridToolbar } from "@mui/x-data-grid";
 import i18n from "i18next";
@@ -11,16 +11,19 @@ import { useApiRequest } from "../../../hooks/useApiRequest.ts";
 import { muiLocales } from "../../../mui.locales.ts";
 import { TablePaginationActions } from "../../overview/boreholeTable/TablePaginationActions.tsx";
 import { quickFilterStyles } from "./quickfilterStyles.ts";
-import { SettingsHeaderContext } from "./settingsHeaderContext.tsx";
 import { useSharedTableColumns } from "./useSharedTableColumns.tsx";
 
-export const UserDetail = () => {
+interface UserDetailProps {
+  user: User | null;
+  setUser: (user: User | null) => void;
+}
+
+export const UserDetail: FC<UserDetailProps> = ({ user, setUser }) => {
   const { id } = useParams<{ id: string }>();
   const { t } = useTranslation();
-  const [user, setUser] = useState<User>();
   const [userWorkgroups, setUserWorkgroups] = useState<object[]>();
   const { callApiWithErrorHandling, callApiWithRollback } = useApiRequest();
-  const { setHeaderTitle, setChipContent } = useContext(SettingsHeaderContext);
+  const history = useHistory();
   const { statusColumn, deleteColumn } = useSharedTableColumns();
   const [filterModel, setFilterModel] = useState<GridFilterModel>();
   const handleFilterModelChange = useCallback((newModel: GridFilterModel) => setFilterModel(newModel), []);
@@ -45,17 +48,20 @@ export const UserDetail = () => {
   useEffect(() => {
     const getUser = async () => {
       const user: User = await callApiWithErrorHandling(fetchUser, [parseInt(id)]);
-      setUser(user);
-      setHeaderTitle(user.name);
-      setChipContent("user");
+      if (!user) {
+        history.push("/setting#users");
+      } else {
+        setUser(user);
 
-      // Get the transformed array of unique workgroups with roles
-      setUserWorkgroups(getUniqueWorkgroups(user));
+        // Get the transformed array of unique workgroups with roles
+        setUserWorkgroups(getUniqueWorkgroups(user));
+      }
     };
     getUser();
-  }, [callApiWithErrorHandling, getUniqueWorkgroups, id, setChipContent, setHeaderTitle]);
+  }, [callApiWithErrorHandling, getUniqueWorkgroups, history, id, setUser]);
 
   if (!user) return;
+  const isDisabled = user.isDisabled;
 
   const renderRoleChips = (params: GridRenderCellParams<object[]>) => {
     return (
@@ -103,10 +109,17 @@ export const UserDetail = () => {
     }
   };
 
+  const disabledStyles = {
+    cursor: isDisabled ? "default" : "pointer",
+    "& .MuiDataGrid-row:hover": { backgroundColor: isDisabled && "rgba(0,0,0,0)" },
+    "& .MuiDataGrid-columnHeader": { cursor: isDisabled ? "default" : "pointer" },
+  };
+
   return (
     <Stack
       sx={{
         height: "100%",
+        opacity: isDisabled ? "50%" : "100%",
         p: 5,
         overflowY: "auto",
         backgroundColor: theme.palette.background.lightgrey,
@@ -115,7 +128,12 @@ export const UserDetail = () => {
         <CardHeader title={t("general")} sx={{ p: 4, pb: 3 }} titleTypographyProps={{ variant: "h5" }} />
         <CardContent sx={{ pt: 4, px: 3 }}>
           <Stack direction={"row"} alignItems={"center"}>
-            <Checkbox checked={user.isAdmin} onChange={handleCheckboxChange} data-cy="is-user-admin-checkbox" />
+            <Checkbox
+              checked={user.isAdmin}
+              onChange={handleCheckboxChange}
+              data-cy="is-user-admin-checkbox"
+              disabled={isDisabled}
+            />
             <Typography>Admin</Typography>
           </Stack>
         </CardContent>
@@ -125,7 +143,11 @@ export const UserDetail = () => {
         <CardContent sx={{ pt: 4, px: 3 }}>
           {userWorkgroups && (
             <DataGrid
-              sx={{ border: "none !important", ...quickFilterStyles }}
+              sx={{
+                border: "none !important",
+                ...quickFilterStyles,
+                ...disabledStyles,
+              }}
               data-cy="user-workgroups-table"
               columnHeaderHeight={44}
               rowHeight={44}
@@ -152,6 +174,8 @@ export const UserDetail = () => {
               disableRowSelectionOnClick
               hideFooterSelectedRowCount
               disableColumnFilter
+              disableColumnSorting={isDisabled}
+              disableColumnResize={isDisabled}
               disableColumnMenu={true}
               disableDensitySelector
               filterModel={filterModel}