+++ src/client/src/pages/settings/admin/userTable.tsx
@@ -1,43 +1,53 @@
-import { ChangeEvent, useCallback, useContext, useEffect, useState } from "react";
+import { ChangeEvent, FC, MouseEvent, useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useHistory } from "react-router-dom";
-import { Checkbox, Chip, Stack, Tooltip } from "@mui/material";
+import { Button, Checkbox, Chip, Stack, Tooltip } from "@mui/material";
 import {
   DataGrid,
   GridColDef,
   GridEventListener,
   GridFilterModel,
   GridRenderCellParams,
+  GridRowParams,
   GridToolbar,
 } from "@mui/x-data-grid";
+import { Trash2 } from "lucide-react";
 import { User, WorkgroupRole } from "../../../api/apiInterfaces.ts";
 import { fetchUsers, updateUser } from "../../../api/user.ts";
+import { theme } from "../../../AppTheme.ts";
 import { useApiRequest } from "../../../hooks/useApiRequest.ts";
 import { muiLocales } from "../../../mui.locales.ts";
 import { TablePaginationActions } from "../../overview/boreholeTable/TablePaginationActions.tsx";
 import { quickFilterStyles } from "./quickfilterStyles.ts";
-import { SettingsHeaderContext } from "./settingsHeaderContext.tsx";
+import { useDeleteUserPrompts } from "./useDeleteUserPrompts.tsx";
 import { useSharedTableColumns } from "./useSharedTableColumns.tsx";
 
-export const UserTable = () => {
+interface UserTableProps {
+  setSelectedUser: (user: User | null) => void;
+  users: User[];
+  setUsers: (users: User[]) => void;
+}
+
+export const UserTable: FC<UserTableProps> = ({ setSelectedUser, users, setUsers }) => {
   const { t, i18n } = useTranslation();
-  const [users, setUsers] = useState<User[]>([]);
   const [filterModel, setFilterModel] = useState<GridFilterModel>();
+  const [isLoading, setIsLoading] = useState(true);
   const history = useHistory();
-  const { setHeaderTitle, setChipContent } = useContext(SettingsHeaderContext);
   const { callApiWithErrorHandling, callApiWithRollback } = useApiRequest();
-  const { statusColumn, deleteColumn } = useSharedTableColumns();
+  const { statusColumn } = useSharedTableColumns();
+  const { showDeleteWarning } = useDeleteUserPrompts(setSelectedUser, users, setUsers);
   const handleFilterModelChange = useCallback((newModel: GridFilterModel) => setFilterModel(newModel), []);
 
   useEffect(() => {
+    setIsLoading(true);
     const getUsers = async () => {
       const users: User[] = await callApiWithErrorHandling(fetchUsers, []);
       setUsers(users);
+      setIsLoading(false);
     };
     getUsers();
-    setHeaderTitle("settings");
-    setChipContent("");
-  }, [callApiWithErrorHandling, setChipContent, setHeaderTitle, t]);
+    setSelectedUser(null);
+  }, [callApiWithErrorHandling, setSelectedUser, setUsers, t]);
 
   const renderCellCheckbox = (params: GridRenderCellParams) => {
     const handleCheckBoxClick = async (event: ChangeEvent<HTMLInputElement>, id: number) => {
@@ -58,11 +68,34 @@ export const UserTable = () => {
     return (
       <Checkbox
         checked={params.value}
+        disabled={params.row.isDisabled}
         onChange={event => handleCheckBoxClick(event, params.id as number)}
         onClick={event => event.stopPropagation()}
       />
     );
   };
+
+  const renderCellDelete = (params: GridRenderCellParams) => {
+    const handleDeleteUser = (event: MouseEvent<HTMLButtonElement>, id: number) => {
+      event.stopPropagation();
+      const user = users.find(user => user.id === id);
+      if (!user) return;
+      setSelectedUser(user);
+      showDeleteWarning(user);
+    };
+
+    return (
+      <Button
+        variant="outlined"
+        key={params.row.id}
+        data-cy={`delete-user-${params.row.firstName}`}
+        onClick={event => handleDeleteUser(event, params.id as number)}
+        sx={{ p: 0.5 }}>
+        <Trash2 color={theme.palette.primary.main} />
+      </Button>
+    );
+  };
+
   const renderWorkgroupChips = (params: GridRenderCellParams<WorkgroupRole[]>) => {
     const averageCharacterWidth = 7.5;
     const chipPadding = 16;
@@ -139,17 +172,37 @@ export const UserTable = () => {
       width: 320,
       renderCell: renderWorkgroupChips,
     },
-    deleteColumn,
+    {
+      field: "delete",
+      headerName: "",
+      width: 32,
+      resizable: false,
+      sortable: false,
+      filterable: false,
+      disableColumnMenu: true,
+      disableReorder: true,
+      disableExport: true,
+      renderCell: renderCellDelete,
+    },
   ];
+
+  const getRowClassName = (params: GridRowParams) => {
+    let css = "";
+    if (params.row.isDisabled) {
+      css = "disabled-row ";
+    }
+    return css;
+  };
 
   return (
     <DataGrid
       sx={{ border: "none !important", ...quickFilterStyles }}
       data-cy="users-table"
       columnHeaderHeight={44}
+      getRowClassName={getRowClassName}
       rowHeight={44}
       sortingOrder={["asc", "desc"]}
-      loading={!users?.length}
+      loading={isLoading}
       onRowClick={handleRowClick}
       rowCount={users?.length}
       rows={users}