+++ src/client/src/pages/settings/settingsPage.tsx
@@ -1,27 +1,29 @@
-import { useMemo } from "react";
+import { useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 import { Route, Switch } from "react-router-dom";
 import { Stack } from "@mui/material";
 import { ReduxRootState, User } from "../../api-lib/ReduxStateInterfaces.ts";
+import { User as UserV2 } from "../../api/apiInterfaces.ts";
 import { theme } from "../../AppTheme.ts";
 import { useAuth } from "../../auth/useBdmsAuth.tsx";
 import { TabPanel } from "../../components/tabs/tabPanel.tsx";
-import { SettingsHeader } from "../detail/settingsHeader.tsx";
 import AboutSettings from "./aboutSettings";
 import AdminSettings from "./admin/adminSettings";
-import { SettingsHeaderProvider } from "./admin/settingsHeaderContext.tsx";
 import { UserDetail } from "./admin/userDetail.tsx";
 import { UserTable } from "./admin/userTable.tsx";
 import EditorSettings from "./editorSettings.tsx";
+import { SettingsHeader } from "./settingsHeader.tsx";
 import TermSettings from "./termSettings";
 
 export const SettingsPage = () => {
   const auth = useAuth();
   const { t } = useTranslation();
-  const user: User = useSelector((state: ReduxRootState) => state.core_user);
+  const currentUser: User = useSelector((state: ReduxRootState) => state.core_user);
+  const [selectedUser, setSelectedUser] = useState<UserV2 | null>(null);
+  const [users, setUsers] = useState<UserV2[]>([]);
 
-  const isAdminUser = user.data.admin;
+  const isAdminUser = currentUser.data.admin;
   const isAnonymousUser = auth.anonymousModeEnabled;
 
   const tabs = useMemo(() => {
@@ -33,17 +35,26 @@ export const SettingsPage = () => {
     }
     if (isAdminUser) {
       tabsArray.unshift({ label: t("workgroups"), hash: "workgroups", component: <AdminSettings /> });
-      tabsArray.unshift({ label: t("users"), hash: "users", component: <UserTable /> });
+      tabsArray.unshift({
+        label: t("users"),
+        hash: "users",
+        component: <UserTable setSelectedUser={setSelectedUser} users={users} setUsers={setUsers} />,
+      });
     }
 
     return tabsArray;
-  }, [isAdminUser, isAnonymousUser, t]);
+  }, [isAdminUser, isAnonymousUser, t, users]);
 
   return (
-    <SettingsHeaderProvider>
-      <SettingsHeader />
+    <>
+      <SettingsHeader selectedUser={selectedUser} setSelectedUser={setSelectedUser} users={users} setUsers={setUsers} />
       <Switch>
-        <Route exact={false} key={4} path={"/setting/user/:id"} render={() => <UserDetail />} />
+        <Route
+          exact={false}
+          key={4}
+          path={"/setting/user/:id"}
+          render={() => <UserDetail user={selectedUser} setUser={setSelectedUser} />}
+        />
         <Route
           exact={false}
           key={4}
@@ -61,6 +72,6 @@ export const SettingsPage = () => {
           )}
         />
       </Switch>
-    </SettingsHeaderProvider>
+    </>
   );
 };